Code Modernization: Replace usage of strpos() with str_contains()
Fixes #32863

str_contains() was introduced in PHP 8.0 to perform a case-sensitive check indicating if the string to search in (haystack) contains the given substring (needle).

WordPress core includes a polyfill for str_contains() on PHP < 8.0 as of WordPress 5.9.

This commit replaces strpos() with str_contains(), making the code more readable and consistent, as well as better aligned with modern development practices.

Related core ticket: https://core.trac.wordpress.org/ticket/58206

Note: this does not make changes in a few places:
- The Analyzer, Autoloader, Changelogger, WAF packages
- The VaultPress plugin since it can still be used on older versions of WordPress, where the str_starts_with() shim did not exist.