RNMobile: Ensure text is always visible within Contact Info block (#33873)
Fixes an issue with the Contact Info block within the app, in which typed text can sometimes be invisible.

The issue occurs when a block based theme is active with a background colour that contrasts with the device's light/dark mode. For example, if a theme's background is light while the device has dark mode enabled, the text will be white on white.