Declare PHP version requirement in all Composer packages
This makes for a better experience for third parties using the packages,
as they can see on Packagist which PHP versions the package is supposed
to support.

Also this is a good excuse to make a major version bump for all the
packages following #34126 which removed testing with PHP 5.6 and allows
use of 7.0+ features going forward.

And since we're making systematic use of declared PHP versions now,
let's lint for that and make use of it to skip phpunit tests where
appropriate.