Publicize: do not display module toggles when not connected
When an admin is not connected to WordPress.com, let's not display the toggle to turn Publicize on or off. This way they would not inadvertently disable a feature that other authors on the site may be using for their own posts.

The invitation to connect your account to WordPress.com remains, so you can take action and toggle the module once you're connected.

Related discussion: p1699929025605869-slack-CDLH4C1UZ