Code Modernization: replace substr with str_starts/ends_with
Fixes #32864

str_starts_with() and str_ends_with() were introduced in PHP 8.0 to perform a case-sensitive check indicating if the string to search in (haystack) begins or ends with the given substring (needle).

WordPress core includes a polyfill for these functions on PHP < 8.0 as of WordPress 5.9.

This commit uses str_starts_with() and str_ends_with() where appropriate.

This aims to make the code more readable and consistent, as well as better aligned with modern development practices.

Note: this doesn't update uses in a few places, like the autoloader since it can be used outside of WordPress.

Matching core change: https://core.trac.wordpress.org/changeset/55990