ci: Automatically tag project releases
This will automatically create tags like "packages/autoloader@3.0.2" or
"plugins/jetpack@12.9" as projects are released.

The logic used is similar to the autotagger: on push to trunk or
prerelease for projects without a release-branch-prefix and on push to
corresponding release branches for projects with a
release-branch-prefix, if the `/changelog/` directory for the project is
empty and the first entry in CHANGELOG.md is a release version number,
it'll create the corresponding tag for the project.