Repo Gardening: assign teams to an issue based on issue labels.
This commit extends the existing automations linking issues to the "One Board" GitHub Project board.

That board includes a "Team" field, that until now has been manually filled by triagers. However, we can aim to automate some of that process and automatically "assign" (I use assign with quotes here since we're using a column in a project board, not GitHub's built-in assignment setup) issues to specific teams based on the labels present on the issue.
This can be especially useful when combined with a tool that allows creating issues with pre-filled labels, like Bugomattic.

Of note, and since this GitHub Action is also available to other GitHub users, I've added an Automattic specificity to the task:

- When in Automattic repos, the task will look for a mapping of GitHub labels <> Team from a file that ships with the action.
- When not in Automattic repos, folks can specify their own mapping via an action input (labels_team_assignments).

A few notes about the mapping structure:

- The structure also supports slack IDs and GitHub Project IDs. Those will become useful in future commits.
- It groups items by feature, not by team. This allows easier maintenance when a feature switches owners; you only need to update the team name, you do not need to move potentially multiple labels around.
- Since we group items by feature, we can specify different Slack channels and project boards for different features. This way, teams working on multiple features can handle each one of them separately.