Slack: extract utilities to own directory
Those functions will be used in multiple tasks, so they can be extracted to avoid code duplication.

Of note, I'm taking this opportunity to move fetching the Slack token to the function that sends the messages out, again to avoid code duplication.
In a follow-up commit, I will remove the slack token fetchs from existing tasks. In practice, this may mean that some tasks will fail later in the process when folks do not specify a token, but I think this is an okay trade-off.