cli: Update test for node 20.10 fixing a bug (#34371)
Until Node 20.10, the callback passed to a stream filter's `.end()`
function would not be called if the stream raises an error. We had a
test checking that behavior.

Now it is called, so the test needs updating. For now we'll introduce a
version check so we don't have to bump the node version for everything.