cli: Avoid node bug in process plumbing
In `listProjectFiles` we first run `git ls-files` and process its
output, then run a second `git ls-files` with different options and
piped to `git check-attr` and process its output.

To keep the promise handling less confusing, we create all those
processes at once and only attach the second pipe's output to the
processor after the first pipe finishes.

But it seems Node has a bug (or at least unexpected behavior) where it
throws away its internal stream buffer for the second pipe's output if
the final process exits before the first pipe finishes.

The most logical (if not the most straightforward to code) way to handle
that is to not even create the second pipe until the first one finishes.