Repo Gardening: assign teams to an issue based on issue labels. (#34313)
* Repo Gardening: assign teams to an issue based on issue labels.

This commit extends the existing automations linking issues to the "One Board" GitHub Project board.

That board includes a "Team" field, that until now has been manually filled by triagers. However, we can aim to automate some of that process and automatically "assign" (I use assign with quotes here since we're using a column in a project board, not GitHub's built-in assignment setup) issues to specific teams based on the labels present on the issue.
This can be especially useful when combined with a tool that allows creating issues with pre-filled labels, like Bugomattic.

Of note, and since this GitHub Action is also available to other GitHub users, I've added an Automattic specificity to the task:

- When in Automattic repos, the task will look for a mapping of GitHub labels <> Team from a file that ships with the action.
- When not in Automattic repos, folks can specify their own mapping via an action input (labels_team_assignments).

A few notes about the mapping structure:

- The structure also supports slack IDs and GitHub Project IDs. Those will become useful in future commits.
- It groups items by feature, not by team. This allows easier maintenance when a feature switches owners; you only need to update the team name, you do not need to move potentially multiple labels around.
- Since we group items by feature, we can specify different Slack channels and project boards for different features. This way, teams working on multiple features can handle each one of them separately.

* Extract functions that parse issue body/comment content

Some of those functions will be used in multiple tasks, so this should avoid duplication.

They will be removed from the tasks where they already exist in a follow-up commit.

* Extract functions fetching labels to own directory

Those functions will be used in multiple tasks. Let's extract them to avoid duplication.
They will be removed from the places where they already exist in a follow-up commit.

* Slack: extract utilities to own directory

Those functions will be used in multiple tasks, so they can be extracted to avoid code duplication.

Of note, I'm taking this opportunity to move fetching the Slack token to the function that sends the messages out, again to avoid code duplication.
In a follow-up commit, I will remove the slack token fetchs from existing tasks. In practice, this may mean that some tasks will fail later in the process when folks do not specify a token, but I think this is an okay trade-off.

* Update imports

* Remove existing functions now extracted

* Extract slack notification for important bugs to its own util

This will be useful so we can use the function for other tasks.

* Update debug comments

* Introduce slack notifications for teams that have a slack channel

In the team <> labels mapping, teams can specify a Slack channel ID. If they do, a notification will be sent to their Slack channel whenever a high priority is flagged and assigned to their team.

* fix reference to getLabels

* Fix typo

* Switch to full URLs for board IDs

It is necessary if we want to support both user and organiation boards.

* Automatically add issue to Project board if specified

When a team uses a specific Project board to keep track of their work, or of the work on a specific feature, they can specify it in the label mapping.
We then attempt to add the issue to the board.

* Fix team assignment variable

* Update debug message

* Update docs

* Add more documentation to the task

* Order list alphabetically.

* Also trigger the task when new issues are being created