ci: Fix lock file up-to-date check (#34554)
We suddenly had the check start failing today due to a ranged
require-dev on codeception in plugins/crm (needed to allow tests to run
on all our PHP versions) combined with an upstream release of the
packages.

Looking at the history of the test, we originally (#18170) did
`composer update --root-reqs` to try to limit the checking to monorepo
packages. But that didn't really work, so in #19807 we changed to a
custom script that explicitly specifies all the relevant monorepo
packages for updating. But we left the `--root-reqs` there, despite it
now having a different meaning (expanding to add root reqs, rather than
limiting to just root reqs).

Looking at it now, the `--root-reqs` isn't really needed. With or
without it'll check monorepo packages, and with or without it'll flag if
what's locked doesn't match what's in composer.json.