cli: Pass only needed dirs to chokidar for rsync --watch
At least on Linux just telling chokidar to watch the plugin dir causes
it to try to watch hundreds of thousands of paths, digging deep inside
`node_modules/` and recursively through `vendor/` and such.

Instead, let's configure chokidar with `depth: 0` and manually tell it
each directory we want it to watch.

That also means refactoring the code that builds the rsync filter into
separate code to collect the list of paths and then code to turn that
into rsync's syntax.

Then we'll also want to skip the prompting for saving the destination
after each rsync, as ideally there should be no user interaction needed
with --watch. Unfortunately I couldn't figure out a method to have ssh
not prompt for a password; if you're trying to rsync-watch to JN you'll
first have to set up a public key manually.