ci: Disable "CI Results" workflow on forks (#34951)
Sometimes someone will create a fork and set it up to auto-update, but
not disable all the workflows that fail without secrets. In this
situation, it seems the CI Results workflow will attempt to run, will
fail, and GitHub will then email the author of the commit about it.

Add a check to that workflow so it will only run from
Automattic/jetpack, not from any other repo.