jetpack: Fix webpack publicPath with JS concatenation (#34969)
In #34923 we switched from setting Jetpack_Block_Assets_Base_Url to
using webpack's automatic `publicPath`. Unfortunately this breaks with
JS concatenation, as done for example by WordPress.com Simple.

So bring back the variable, but do it in a way such that the variable
only gets set once instead of for every block.