Changelogger: Default `--deduplicate` to 0 (#35138)
When set to 1, it looks at the previous changelog entry and removes any
from the current entry that already show there. While that makes sense
in the case of "normal" changes that might have been backported, it
makes much less sense for perennial things like "Updated dependencies".
So probably better if people have to opt in to that behavior.