Admin-UI: Avoid loading l10n too early in admin
When on a plan with anti-spam, the Akismet menu is being added during
the 'plugins_loaded' action, which is before the user has been
determined and so it loads the l10n in the site language for the
plugin's text domain.

Delay that addition until the 'admin_menu' action (at the same time as
when we remove Akismet's copy of the menu) to avoid triggering the l10n
so early.