webpack-config: Add fork-ts-checker-webpack-plugin option
We've decided (pdWQjU-FO-p2) to prefer `@babel/preset-typescript` over
`ts-loader` for dealing with TypeScript in the context of Webpack. For
generating the definition files, we're generally going to use
`fork-ts-checker-webpack-plugin` (although `tsc --emitDeclarationOnly`
is also an option).

To make that as easy as possible for people to work with, let's make it
an option in the shared webpack config.