Sharing Block: automatically add block to templates in block themes
Under a few specific conditions, we want to automatically add a sharing buttons block (and 3 inner blocks, Facebook, X, and Mastodon) to the bottom of the post content.
Site owners can then go to Appearance > Editor if they want to remove or customize that block.

For now, we're trying to automatically add the block right after the post content block, in post and page templates (that includes templates like `page-wide`, `page-no-sidebar`, etc).

Here are the conditions:

- The site must be using a block theme.
- An option (`jetpack_sharing_buttons_auto_add`) must be set. Later on we can add UI to set that option under Jetpack > Settings > Social.
- The sharing module must be deactivated. We wouldn't want to automatically add sharing buttons to the post content twice.
- On WordPress.com Simple, the module is always active. There, we check if sharing posts are enabled on single posts and / or pages, since that matches the template where we'd like to automatically add the block.

Resources:

- https://make.wordpress.org/core/2023/10/15/introducing-block-hooks-for-dynamic-blocks/
- https://github.com/ockham/like-button
- https://core.trac.wordpress.org/ticket/59572
- https://core.trac.wordpress.org/ticket/60126

Internal references:

- p1707216098587689-slack-C065QF2R1B6