Blaze: allow switching off quick action links for specific CPTs (#35730)
Follow-up to #35620

If we move the filter to a bit later in the process, we allow removing the Blaze link from some post types, but not others. In this scenario, in the Jetpack plugin we can turn off the links for posts and pages, while still allowing them for products.