cli: Add `jetpack install --use-uncommitted-composer-lock`
It's a lot faster to run `composer install` rather than `composer update`.
We do the latter for plugins where composer.lock is checked in, since CI
ensures it matches composer.json. We don't for other stuff where
composer.lock isn't checked in because we can't know whether the lock
file is up to date or not.

With this new option, it runs `composer validate` (which is fast) when
an uncommitted composer.lock exists, and if that's happy then it will do
the faster `composer install`.