Scheduled Updates: Always request updates (#36162)
* Scheduled Updates: Always request updates

Sending update requests even if there are no plugins to update, allows WP.com to keep track of that result and display it in the Schedule Updates UI.

* Don't add old_version information

It is not longer needed after D139459-code.

* Use blog token since user information is not available when Crons run

* Don't send anything if there are no plugins to update

The auth check on wpcom gets confused with empty arrays it seems? This works.