ci: Try faking a connection in plugin update test
We've had some cases where problematic code didn't trigger during the
test because it needs the site to be connected.

PR #35911 suggests a way to fake the connection. We can't really use
that directly since we want the code for all plugins, not just
Jetpack-the-plugin, but we can copy-paste the relevant bit of code into
the test script.

Also, while we're at it, let's run each test with a clean copy of the
database.