Storybook: Add support for `@storybook/test-runner` (#35919)
Storybook has a `@storybook/test-runner` package, which can be used to
make sure the stories at least work (and can be used for integration
testing too if such tests are written). This tries adding that package
and running it in CI.

To not have to build all the deps before we can run the test, let's set
`jetpack:src` in storybook's webpack config like we do in our own.
Hopefully that also means we can remove the `extra.dependencies.build`
from composer.json too, which may speed up the builds a little by not
having to have storybook wait until the end.

To make the output from that a bit more friendly, let's also make
storybook development builds not minify.