transport-helper: Fix upgrades from plugins embedding v0.2.0 (#36440)
In v0.2.0, namespaces were changed for classes in the package. This
broke upgrades of plugins embedding the package due to a hook being
registered for one of these classes on startup since the class no longer
existed by the time the hook was called.

v0.2.1 renamed that class back, and added a stub with the versioned name
to try to handle upgrades from v0.2.0. Unfortunately the stub isn't in
the right place: plugins embedding v0.2.0 will have an autoloader that
specifies the namespaced class is in `class-package-version.php` while
the stub is actually in `class-package-version-compat.php`.

This fixes those upgrades by having `class-package-version.php` require
`class-package-version-compat.php` so those autoloaders can still find
the namespaced stub class.