phan: Add stubs for WordPress-defined constants
Unfortunately the php-stubs/wordpress-globals package is
broken-as-designed, because the maintainer doesn't actually want anyone
to use it. Instead he wants people to use PHPStan, which needs a
bootstrap file for constants, which he provides in his PHPStan plugin.

This initial version copies most of the constants from that PHPStan
plugin's bootstrap.