Sharing: clarify 2 options in settings (#36473)
There are currently 2 main options for sharing buttons:

- The legacy sharing buttons, the preferrable option when you use a legacy/classic theme.
- The sharing buttons block, when you use a block-based theme.

We've tried to highlight the 2 options in #34619, but the wording mentioning "below" for a link that in fact lead to the site editor was confusing. This commit updates the wording and introduces a notice instead of a basic paragraph text that may be missed.