autorelease, autotag: Tag and release non-alpha prereleases too (#36482)
It has been a longstanding request to auto-tag and auto-release
non-alpha prerelease versions, particularly betas. Let's make that
happen.

Note that even-numbered "-a.n" versions for Jetpack are considered
alphas, while odd-numbered versions are not (they're what we release to
Atomic). Right now Jetpack doesn't use the auto-tagger or auto-release,
but I see that we've been manually tagging the odd-numbered 'a'
versions in jetpack-production.