SEO Tools: make feature available on non-connected sites.
The module was previously restricted to paid Jetpack plans. When we added the options to customize the different SEO options to wp-admin (in the block editor and in Jetpack settings), we made the module available to all plans. It is still restricted to paid plans on WordPress.com though.

As a result:

- On self-hosted, we do not need to be checking if the feature is available; it always is.
- On WordPress.com, we should be checking if the feature is available.
- On self-hosted, the feature can consequently be available on offline mode.