ci: Run phan with previous-version WP stubs too
Phan can help us detect accidental addition of calls to PHP functions,
classes, and methods added to WordPress core since the supported
previous versions if we supply it the right stub files.

Note for this to really work, though, we'll need to clean up the
existing baselines, since if a file already has an undefined function
being ignored by the baseline then a new undefined function use added in
the same file would be ignored too.