phan: Separate pseudo-projects from "monorepo" (#36467)
We have some non-project directories in the monorepo that have their own
`composer.json` and such. In order for Phan to properly process the
vendor directories of these things, we need to have the CLI run
`composer install` for them and process them with their own phan config.

Doing this wound up needing some refactoring in the cli
`helpers/install.js` to check for composer.lock files in arbitrary
non-project directories.