phan: Rework how stubs are specified
In preparation for adding more specialized stubs, add a way for configs
to specify which they want by keyword rather than having to do file
paths relative to the root.

As part of that, `'is_wordpress' => false` goes away in favor of
overriding the default list of stubs to load.