changelogger: Support symfony/console 7.0
When I tried using stub-generator from Packagist, it broke because it
wasn't actually compatible with symfony/console 7.0 (it needed
specification of a return type on a method). Turns out this is because
changelogger only specified working with up to 6.0, so 7.0 wasn't being
used when it should have been.

Fortunately all the changes are still more-or-less backwards compatible
all the way back to the 3.4 version we need for PHP 7.0 support. We wind
up having to specify the commands' names and descriptions twice now,
but 🤷.