phan: Allow deleting baseline files (#36857)
In many cases it's pointless to keep around the empty baseline files,
ideally they should never be repopulated. Update the tooling to allow
for these to be deleted, in which case they won't be re-created by
`--update-baseline`.

But in case for some reason one does need to be re-created, add a
`--force-update-baseline` option to allow for that.

Also update the GitHub Action checking for baseline changes to handle
these, and also to handle changes in pseudo-project baselines that
wasn't done in #36467.

For the moment I'm only actually removing the empty baselines for stuff
Garage maintains. We can look at the rest in a followup.