Tiled Gallery Block: remove Plans dependency for VIP Support
Follow-up to #13293

Instead of relying on a WordPress.com API request to get the site's plan, and thus to figure out whether the site is on VIP or not, let's rely on the site's constants. That will save us an external request, and will allow to more easily extract the Tiled Gallery block in the future if necessary.