DM Hack Day Project: AI Image Generator
This is the PoC code for the media source version of the 2022 DM Hack
Day project to add an AI Image Generator. It needs ot be used in
conjunction with D145270-code.

It's not really intended to be merged, but as we're exploring other
versions of this, I thought it would be good to get the code up in a
draft PR.