jitm: Fix most Phan issues
The main motivation here is to handle a bit in
`tests/php/test_pre_connection_jitm.php` that's giving different issues
depending on the version of Mockery in use, but I figured I may as well
fix the rest too.

One issue remains. As far as I can tell it's technically a `mixed` from
`$_GET` that's being assumed to be an array in a few places but is
actually a string, but when it's a string it probably should be turned
into a single-element array (as implicitly happens inside the call Phan
flags). I'll leave that for someone who actually knows about this code
to deal with.