push-to-mirrors: Fix slowdown due to `git fetch --filter=tree:0` (#37009)
`git fetch --filter=tree:0` works well when checking the commit messages
for Update-Ref to save downloading a lot of unnecessary data. However,
when pushing, git seems to decide it needs to fetch some portion of that
data anyway, and does so in an inefficient manner.

We can avoid that by making a temporary second `.git` directory and
doing the `git fetch --filter=tree:0` into that instead of into the real
one, so the real one doesn't wind up with whatever weirdness makes git
do the slow data fetch on push.