phan: Fix phpdocs Phan can't parse
A lot are things like `@returns` instead of `@return`, JS syntax like
`@param {int}`, and punctuation like `@return array.`.

One thing particularly of note is that when we have a doc comment for a
hook, then later in the function a closure without a doc comment, Phan
was thinking the hook doc belonged to the closure (and then complained
that none of the `@param`s matched). We can fix that by adding a doc
comment for the closure itself.

WAF was already using some advanced syntax 👍, but they syntax they used
was incompatible with Phan in a few places:
* Phan doesn't like leading/trailing spaces in array shapes, so do like
  `array{int, int}` instead of `array{ int, int }`.
* A few places they did `array<>` rather than `array{}`.
* For some reason they were using `@template` where `@type` was
  intended.