ci: Only do version-number-only phan stub updates monthly
We only really need to update when there are actual changes to the
stubs. So let's only do version-number-only bumps on the first Thursday
of the month.

Any real change in any of the stubs will also update the version numbers
in any other stub files that need it, the Thursday thing is just so they
don't get excessively out of date if none of the stubs ever get updated.