cli: Allow flagging deps as test-only
Our `jetpack dependencies build-order` and `jetpack build` commands will
detect dependency cycles and complain about them. Mostly this has worked
ok so far, we've only had to hack around that in some
`js-packages/i18n-check-webpack-plugin` tests so far.

But now that we have Phan, it turns out there are a lot of other places
where we're missing dependencies but adding them will trigger the cycle
warning.

So let's add an `.extra.dependencies.test-only` key in composer.json,
which the CLI can look at to know that a certain dependency is only for
testing and/or static analysis and so it can ignore them when trying to
decide on the order to build packages in.

We also need to add `ignore-workspace-cycles = true` to `.npmrc` to tell
pnpm not to complain at us about cycles it detects. Our tooling does the
detection better.