packages: Add a bunch of missing deps
The following dependencies are added:

* packages/backup: packages/my-jetpack
* packages/boost-core: packages/connection
* packages/forms: packages/logo, packages/status, packages/sync
* packages/import: packages/sync
* packages/search: packages/sync
* packages/videopress: packages/my-jetpack
* packages/yoast-promo: packages/assets

Also, the following are added as dev-dependencies and flagged as
test-only so they won't result in dependency cycles:

* packages/config: 14 different things it can conditionally load
* packages/connection: packages/licensing, packages/sync
* packages/my-jetpack: packages/search, and mark packages/videopress as test-only too
* packages/status: packages/connection, packages/identity-crisis, packages/plans
* packages/sync: packages/search, packages/waf

Note I haven't fully checked that all these really are test-only with
any non-test uses properly checked with `class_exists` or the like. I do
note that status→connection is only because of the `Jetpack_Options`
class and is _not_ checked, which seems unfortunate and probably should
be fixed somehow at some point.