ci: Avoid suggesting unnecessary baseline updates
When checking for baselines that need updating due to fixed issues, we
ignore the "issue statistics" comment at the top of the file so as to
not make people have to update the baselines unnecessarily.

However, if one project has a real update needed while a different one
has only comments, we're still instructing the author to update both of
them even though only the former is needed.

Let's not do that.